Move switch command into separate file.
